+++ tests/test_functions.py
@@ -1,0 +1,59 @@
+import json
+
+import pytest
+from shillelagh.backends.apsw.db import connect
+from shillelagh.exceptions import ProgrammingError
+from shillelagh.functions import get_metadata
+
+from .fakes import FakeAdapter
+from .fakes import FakeEntryPoint
+
+
+def test_sleep_from_sql(mocker):
+    sleep = mocker.patch("time.sleep")
+    entry_points = [FakeEntryPoint("dummy", FakeAdapter)]
+    mocker.patch(
+        "shillelagh.backends.apsw.db.iter_entry_points",
+        return_value=entry_points,
+    )
+    connection = connect(":memory:", ["dummy"], isolation_level="IMMEDIATE")
+    cursor = connection.cursor()
+    cursor.execute("SELECT sleep(5)")
+
+    sleep.assert_called_with(5)
+
+
+def test_get_metadata(mocker):
+    entry_points = [FakeEntryPoint("dummy", FakeAdapter)]
+    mocker.patch(
+        "shillelagh.functions.iter_entry_points",
+        return_value=entry_points,
+    )
+    assert (
+        get_metadata(
+            {"dummy": ("foo",), "other": ("bar",)},
+            {"dummy": {"key": "value"}, "other": {"one": "two"}},
+            "dummy://",
+        )
+        == '{"extra": {}, "adapter": "FakeAdapter"}'
+    )
+
+    with pytest.raises(ProgrammingError) as excinfo:
+        get_metadata({}, {}, "unsupported://")
+    assert str(excinfo.value) == "Unsupported table: unsupported://"
+
+
+def test_get_metadata_from_sql(mocker):
+    mocker.patch(
+        "shillelagh.functions.get_metadata",
+        return_value=json.dumps({"hello": "world"}),
+    )
+    entry_points = [FakeEntryPoint("dummy", FakeAdapter)]
+    mocker.patch(
+        "shillelagh.backends.apsw.db.iter_entry_points",
+        return_value=entry_points,
+    )
+    connection = connect(":memory:", ["dummy"], isolation_level="IMMEDIATE")
+    cursor = connection.cursor()
+    cursor.execute('SELECT get_metadata("dummy://")')
+    assert cursor.fetchall() == [('{"hello": "world"}',)]