+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Version 0.7.0 - 2021-MM-DD
 
 - Add support for DML to the Google Sheets adapter
 - Improved many small bugs in the type conversion system
+- Add ``sleep`` and ``get_metadata`` functions
 
 Version 0.6.1 - 2021-06-22
 ==========================